requirements.txt
# Flask and extensions
flask[async]
flask-cors
python-dotenv
pyjwt
asgiref  # For async support
aioflask  # Add explicit async Flask support
hypercorn  # ASGI server for Flask

# Database
supabase

# OAuth and API
requests==2.31.0
PyJWT==2.8.0

# Notion SDK
notion-client==1.0.0

# Utils
pydantic>=1.8.2

# Additional dependencies
cryptography

# Add Supabase client
supabase>=2.11.0  # Latest version as of now

# Vapi SDK
vapi_server_sdk 